Refractores Attendee component
